Finished working code for LED dashboard. Includes c#, arduino, and notes.
This commit is a complete working system. I have included my notepad
notes and the arduino code. It could all use some polish. Also I should
add a reset button to the GUI. Where I would like to go from here is to
encapsulate all of it in classes and create libraries for the design. In
addition. I would like to create a working virtual demo of the actual
LED for testing without having to hook up an actual LED array every
time.